wrappers: converge 3 platform trees via full drift-gate + honest re-adopt (task 21)
Collapse the desktop/Android/iOS generated wrapper sets onto one enforced source
of truth: the honest (task 11) generator. Re-adopt the accumulated drift and add
a per-platform CI drift-gate so committed can never again diverge from a fresh
regen. This is task 06's convergence *goal* achieved via enforcement; the
physical :kanama-common-api KMP module move stays a staged follow-up (the
drift-gate already makes cross-platform wrapper drift structurally impossible).

Re-adoption (from the now-honest generator):
- Desktop: regenerated + re-adopted 985 in-API wrappers. Real drift was ~985
  files, not the ~273 estimate — task 11's honesty (enum consts, fromHandle
  factories, composite defaults, openness) compounded pre-existing drift and the
  old inline-vs-named method-bind-hash style. KDoc preserved by splicing the
  committed per-member docs back onto regenerated code (sync_kdoc still owns docs).
- iOS: adopted 8 audited-subset diffs (composite defaults); hand-written glue
  (Engine/SceneTree/Tween/ProjectSettings/...) untouched.
- Android: no separate committed tree — prepareAndroidKanamaSources copies the
  desktop sources through the PanamaPort remap, so desktop adoption covers it.
- 46 desktop classes are deliberately hand-shaped (facades like DirAccess/
  FileAccess/RefCounted/Resource, ergonomic helpers on Node/Viewport/..., custom
  defaults on TabBar/Node3D) and are the only drift-gate exemptions.

Full drift-gate (scripts/check_wrapper_generator.py):
- New check_full_drift_gate regenerates the ENTIRE committed set per platform and
  fails on any divergence (behavior-comparable; KDoc stripped). Replaces the
  representative-fixture-only coverage. Proven to catch an injected hand-edit.
- comparable_source now strips single-line generated KDoc too, so old-style short
  docs don't false-fail.
- Exemptions frozen in DESKTOP_HANDSHAPED/IOS_HANDSHAPED; collision singletons
  excluded via IOS_HANDWRITTEN_COLLISION_CLASSES.

Two example_project consumer edits reflect honest API (int const -> Long;
joypad device -> Int). Docs refreshed: wrapper-maintenance drift-gate section,
commonMain-design + coverage-tracker status, api-coverage report.

Validated: check_wrapper_generator.py (desktop 985 / iOS 242) green; ./gradlew
jar green; compileKotlinIosArm64 green; Android plugin AAR compile adds 0 errors
vs main under the PanamaPort remap (all pre-existing errors are stale demo KSP);
mkdocs build --strict green. Device gates (iOS device gate + Android Pixel smoke)
pending — no physical device currently attached.

+++ example_project/WrapperConvenienceProbe.kt
@@ -331,7 +331,7 @@ class WrapperConvenienceProbe(val godotObject: MemorySegment) {
 	}
 
 	fun constantsAndFactories(): Camera3D {
-		val invalidCellItem: Int = GridMap.INVALID_CELL_ITEM
+		val invalidCellItem: Long = GridMap.INVALID_CELL_ITEM
 		val bodyAxis: Long = PhysicsBody3D.BODY_AXIS_ANGULAR_X
 
 		return Camera3D.create()
